+++ Maze_Lib_Test/TestA_06.cpp
@@ -29,7 +29,7 @@ int TestA_06()
     TEST_ASSERT(Maze_ERROR_POINTER == lR0.SetMazeName(NULL));
     TEST_ASSERT(Maze_ERROR_ENUM == lR0.SetMode(static_cast<Maze::Race::Mode>(4)));
     TEST_ASSERT(Maze_ERROR_MAX == lR0.SetRunnerCount(4));
-    TEST_ASSERT(Maze_ERROR_MIN == lR0.SetRunnerCount(static_cast<unsigned int>(0)));
+    TEST_ASSERT(Maze_ERROR_MIN == lR0.SetRunnerCount(0));
     TEST_ASSERT(Maze_ERROR_MAX == lR0.SetSlow(1001));
 
     TEST_ASSERT(Maze_ERROR_POINTER == lR0.Property_Set(NULL, "Value"));
